requirements.txt
# Core dependencies
slack-sdk==3.23.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-vision==3.4.4
google-cloud-videointelligence==2.21.0
google-cloud-storage==2.10.0

# Web framework for webhooks
flask==2.3.3
gunicorn==21.2.0

# Data processing
pandas==2.1.3
python-dateutil==2.8.2

# Configuration and utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0

# Logging and monitoring
structlog==23.2.0
colorama==0.4.6

# Database for tracking
sqlite3

# File handling
Pillow==10.1.0
python-magic==0.4.27

# Async support
aiohttp==3.9.1
asyncio-throttle==1.0.2
